Move files to Streamlist folder
