+++ frontend/src/streamlist/DeleteStreamDialogSummoner.jsx
