+++ frontend/src/streamlist/__tests__/DeleteStreamDialogSummoner.test.jsx
